SFDS/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.2
distlib==0.3.8
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
identify==2.5.33
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.4
nodeenv==1.8.0
passlib==1.7.4
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
